ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
spacy==3.7.2
scikit-learn==1.3.1
pydantic==2.4.2
python-multipart==0.0.6
PyMuPDF==1.23.6
python-docx==1.0.1
pytest==8.3.5
